change to uint54_t to avoid overflow

+++ nanopond.c
@@ -319,9 +319,9 @@ static const uintptr_t BITS_IN_FOURBIT_WORD[16] = { 0,1,1,2,1,2,2,3,1,2,2,3,2,3,
 
 /*Set up instruction counting*/
 #ifdef USE_PTHREADS_COUNT
-int p_instruction_counts[USE_PTHREADS_COUNT];
+uint64_t p_instruction_counts[USE_PTHREADS_COUNT];
 #else
-int s_instruction_count = 0;
+uint64_t s_instruction_count = 0;
 #endif
 
 /**
@@ -1108,7 +1108,7 @@ int main()
 
 
     #ifdef USE_PTHREADS_COUNT
-    int total_insts = 0;
+    uint64_t total_insts = 0;
     for (int i=0;i<USE_PTHREADS_COUNT;i++) {
         printf("Thread %d executed %d instructions\n",i,p_instruction_counts[i]);
         total_insts+=p_instruction_counts[i];
